Update README.rst
This article will aim to shed more light on what goes on behind the scenes after you type this simple but important request.
Let me start with ‘what is a browser’? A browser is an app you use to access the web. Examples are : Google chrome, Firefox. Microsoft Edge, Safari etc.
When you type www.google.com into a web browser and press the enter key, a lot of processes happen in those milliseconds before your search results are displayed.
The first thing that happens is your computer sends a request to the domain name system (DNS) server. The DNS server is an address book for all domain names on the web. This DNS now sends back the IP address of www.google.com
The second thing that happens is your computer connects with the server through the IP address. This connection is known as the Transmission control protocol (TCP) through an Internet protocol(IP).
You could call this process a ‘handshake’
Your computer, if it's behind a firewall, checks to see if this request is permitted before the connection is made.
After the connection, your browser sends a request to the web page using an encryption protocol (SSL) or transport layer security (TLS) to encrypt the data shared between your computer and the server. This is the reason for the ‘s’ in https (s) means secure.
Tech companies such as Google who receive a lot of traffic have several servers and load balncers that help to manage the traffic by sending requests to a specific server via a load balancer who sends it back to the server depending on the algorithm of the load balancer.
This response from the load balancer to the server back to your browser(response) includes files such as HTML, CSS, Java script that makes the google homepage.
The web server may request a database server to get data and send it back in response. Finally, the browser will render the page and display it to you.